Register models with the Django admin
During development this will allow us to easily inspect and manipulate the state
of the database and demo backend features before the UI is complete.

We added `django.contrib.gis` to the list of installed apps so that our Point
geometry fields automatically render on OpenLayers maps in the admin.