Adjust login endpoint to prevent creating a token
django-rest-auth creates a token in passing via its LoginView class,
along with creating a session login. However, we'd prefer it only to
begin a session login and to have the user generate auth tokens via the
profile UI if desired.

This commit prevents the form login from creating a token in passing by
replacing the call to `super`/`post` with calls to Django's built in
authenticate and login methods. It also contains adjustments to the
client's Redux actions to work with the new login response.