Remove containers dir
The original React source had a `components` dir and a `containers` dir,
which is an idiomatic distinction meant to map on to a distinction
between React components which are connected to Redux and React
components which are not.

However, the original source files often didn't adhere to that
distinction, and our typical team design pattern is to use Redux state
pretty freely.

To reduce the dev environment's complexity, this commit

- moves the last five remaining files from `containers/` into
`components/`
- updates the import paths to work with the new file locations
- deletes the `containers/` dir