Add corejs library for use in IE11 polyfills
Add corejs library which enables importing polyfills for IE11 on a
method-by-method basis, and to adjust which polyfills are imported while
the development server is running.

I chose this solution as an alternative to using `@babel/polyfill` which
unfortunately caused a 10x increase in the time it took for the client
dev server to start, seemingly because it was parsing all the code in
node_modules for every polyfill.