Make facility data in fixtures consistent
We were previously creating a `Facility` for every `FacilityListItem` but we
were only creating a `FacilityMatch` for a random third of the items and not
setting the status of the `FacilityListItem`s correctly. In this commit we have
updated the `makefixtures` command and the affected fixtures so that the data
is consistent.