Add endpoint for fetching environment variables
Previously, variables could only be injected into the
front-end at compile time. Since we'd like to deploy
the same code to different environments and be able to
plug different third party services in for each of them,
we need a way to load variables at run time.

This new endpoint reads all variables prefixed with
REACT_APP_ from the environment and returns them at the
/web/environment.js endpoint. This is then loaded in to
the front-end before rendering, and the values added to
a window.ENVIRONMENT variable.