Render index via Django, inject variables
Previously, the rendering of index.html was handled at `/`
by WhiteNoise, thanks to WHITENOISE_INDEX_FILE being True.
With it set to False (by default), `/` is handled by oar.urls,
which renders it via the new app web.views.index.

When the Django process starts up, it assembles all environment
variables starting with REACT_APP_ and collects them in a
`context` dict. This is injected in to the index.html template
before being returned to the client. This also includes an
`ENVIRONMENT` value for the current environment.

Since the front-end is a Single Page App with its own routing,
the user may directly request a front-end route from the server,
e.g. `/contribute`. To allow for this, we add a custom handler
for 404s, and set it to render the same template that index
does. This will pass the route on to the index file, allowing it
to handle it.