Render index via Django, inject variables
Previously, `/` was handled by Django-SPA as it automatically
uses any index.html in a directory as the root file. Since we
want to render this via a view (so we can inject variables in
it), we rename the build output file to be main.html instead,
which makes it not render via Django-SPA automatically.

When the Django process starts up, it assembles all environment
variables starting with REACT_APP_ and collects them in a
`context` dict. This is injected in to the main.html template
before being returned to the client. This also includes an
`ENVIRONMENT` value for the current environment.

Since the front-end is a Single Page App with its own routing,
the user may directly request a front-end route from the server,
e.g. `/contribute`. To allow for this, we add a custom handler
for 404s, and set it to render the same template that index
does. This will pass the route on to the index file, allowing it
to handle it.