Provision separate containers for ECS and Batch
AWS Batch does not allow us to redefine the `ENTRYPOINT` of a container at
runtime so we need to create separate images for the web application and batch
processing. We are opting to not rename the `openapparelregistry` repo/image to
`openapparelregistry-app` to avoid having terraform destroy the existing
repository while this is change is being tested in a shared staging environment.