Set item.facility in current match process
If the matching process has created a new facility from the list item or
if the matching process has found a facility match with 100% confidence,
set `item.facility = facility`.

Note that in the current implementation creating a new facility or
making an automatic 100% match are the only two possible cases, but
we'll be updating this as we implement the new match process.