Use BATCH_MODE env var to disable checking instance metadata
When the container is run as a normal Django web app process in Staging or
Production we check the instance metadata to properly set the `ALLOWED_HOSTS`
Django setting. This instance metadata check fails when the container is run in
AWS Batch so we have added a the `BATCH_MODE` environment variable which acts as
a switch to disable the instance metadata check.