Submit AWS Batch jobs when CSV is uploaded to staging or production
Jobs are only submitted in staging an production because AWS Batch cannot run in
development.

This code assumes that we have a single job queue and we will always want to run
the latest version of the job definition. This may change in the future if we
expand our processing but for now those assumptions make for simpler code.

We record the job IDs in the `FacilityListItem.processing_results`. It is a bit
redundant to log the same ID on all the line items but it is consistent with the
fact that all processing is done at the item level and allows us to display the
entire processing history of an item without needing to collate detail from the
item and its parent list.

There is no way to resubmit jobs if the process fails. In this initial version
the user can work around that type failure by reuploading the same CSV and using
the "replaces" feature.