Add a dedupe-based facility list matching function
Experiments with a pre-trained model revealed that the IDs of the canonical
items used while training are returned as match results. This means that the
model must be trained using actual data from the facilities table.

The training.json file that is read in during model training was created by
running through an active learning session using the oar_gazetteer_example.py
research script against the 42 example public facility CSVs used to populate the
previous version of the application.

We have updated the Vagrantfile and docker-compose.yml file to ensure that our
development environment is provisioned with 2 cores. This avoid a bug uncovered
during our gazetteer research that only appears when there is a single core
available. See https://github.com/dedupeio/dedupe/pull/734

Dedupe is designed to work with an entire list of items, not one at a time, so
we have added a temporary "test" action to the `batch_process` management
command to exercise the matching methods against fixture data and uploaded CSVs.

We added the Unidecode PyPI package following the lead of the dedupe examples.

    Unidecode provides [a function that] takes Unicode data and tries to
    represent it in ASCII characters (i.e., the universally displayable
    characters between 0x00 and 0x7F), where the compromises taken when mapping
    between two character sets are chosen to be near what a human with a US
    keyboard would choose.