Skip Facility List CSV Header Row on Ingest
Previously each line in the Facility List CSV would be added
as a Facility List Item to the database, including the header
row.

To make it so that only true data rows are added, we now skip
the first line of the uploaded CSV file. The added facility
list item's `row_index` field still starts with 0 though, to
help make AWS Batch processing easier, which uses a 0-indexed
scheme.