Rename "Organization" → "Contributor"
Since individuals such as researchers, students, or journalists
may contribute to the site, in addition to organizations such
as a brand, retailer, manufacturer, etc., we switch to using
the more inclusive term "Contributor" throughout the code.

This also brings the back-end terminology in line with what
the UI already uses.

organization → contributor
org → contrib