Upgrade react-scripts, eslint, related items
This more recent version of react-scripts includes securty
updates, but also requires newer versions of eslint, as well
as the new browserslist key in package.json, which are currently
filled with default values.

The very latest version of react-script, 2.1.5, has a bug
that causes styles to not load correctly in the development
envrionment facebook/create-react-app#6399, so we use the
most recent version without that bug, namely 2.1.3.