Update batch_process management command to use new matcher
The new match process is designed to process the full list at once, and does so
efficiently. We have refactored the `batch_process` code to separate "item"
actions from "list" actions. "match" is the only a "list" action at the moment,
and possibly ever, so we did not generalize the calling of the function that
does the actual work.

We moved the handling of the `parse` and `geocode` actions to a `process_items`
method to keep the length of the `handle` method under control.