Add gazetteer_threshold arg to matcher and handle no match case
Adding unit tests uncovered two issues.

When no rows match the canonical list a `BlockingError` is raised which would
crash the match method. We resolved this by catching the exception and returning
an empty list of matches along with a flag value indicating that there were no
matches.

Matches were being found intermittently because the calculated threshold was
sometimes too high. We addressed this by setting the threshold to a value passed
into the `match_facility_list_items` method, which has the same default value of
0.5 that is used by the `dedupe.Gazetter.match` method.