Extract common list labeling functionality
The same technique was used in multiple cases for updating
a list of query string parameters with labels from another
dataset for contributors, contributorTypes, and countries.
This functionality is needed elsewhere, when updating all
filters, so it is extracted from here into a utility file,
and the functions rewritten to be more reusable.