Use correct labels for contributors on app route
Previously, the labels shown in the filter selection box
would match the contributor name only if UPDATE_ALL_FILTERS
fired before COMPLETE_FETCH_CONTRIBUTOR_OPTIONS. The former
set the list of active filters from the query string, and
the latter updated that list with labels after fetching
the data from the API.

In cases when COMPLETE_FETCH_CONTRIBUTOR_OPTIONS fired
before UPDATE_ALL_FILTERS, such as loading the home page,
then going to "My Lists", then selecting "My Facilities",
the filter selection box would show the contributor's ID,
not the label.

By matching the query string filters to the list of saved
contributors in the state before firing UPDATE_ALL_FILTERS,
we ensure that correct label is used in the case when it
fires after COMPLETE_FETCH_CONTRIBUTOR_OPTIONS.

This ensures that the filter selection box has the correct
value when the page is loaded for the first time (e.g. full
page refresh with the query string params), or if it is
navigated to from within the app (e.g. clicking "My Facilities").