Allow a list item with no geocode results to proceed to matching
The dedupe matcher may be able to match a list item to an existing, geocoded
facility even if the geocoder returns no results. To allow for this we add a new
GEOCODED_NO_RESULTS status.

We add a `result_count` to the response from the geocoder so we can easily
distinguish a successful geocode from one that did not fail, but also did not
produce any results.

Items that cannot be geocoded and also fail to be matched end up in the
`ERROR_MATCHING` status because we cannot create a new Facility without a
lat/lng.