Replace custom Translate component w/ standard UI
Recent changes to the app broke the previous implementation of Google
Translate using the Translate.jsx component. Further investigation
revealed that the way the Translate component had been set up was not
really in harmony with how Google Translate is supposed to be set up --
which increased its surface area for bugs and made it incompatible with
existing documentation like

https://www.w3schools.com/howto/howto_google_translate.asp

To remedy this, we drop the Translate component and use the standard
method for setting up Google Translate: giving Google Translate a div on
which to create its own UI that's managed outside of React. This
circumvents having to manually set the hash, call window.reload, and
sync the cookies. It also puts the Google Translate usage back into
harmony with the documentation that is available online.

Note that it turns out that Google Translate for websites seems to be
deprecated, so although it is currently working here we should probably
be prepared for it to break at any point in the future.