Add command to bulk create OAR managed accounts
Adds a command that takes a list of contributor names,
a contributor type, and a password, and creates user and
contributor accounts with those details and some defaults.

These accounts will be added by OAR initially, but may be
transferred to different owners in the future.

Validates that the specified contributor type is one of
the allowed values, and that the password meets regular
app requirements.