Implement server side paging for facility list items
We were running into gateway timeout issues when attempting to load all items
for a large list (thousands of rows)

To solve this we have added a separate endpoint for fetching list items and
implemented paging based on the built-in Django REST Framework paging support.

We were able to reuse the bulk of the client side paging logic but we now fetch
new data when the page or page size is changed rather than just getting a slice
of the full data in memory.