Increase the max length of FacilityListItem.geocoded_address
We were running into crashes in the geocoding process because we were trying to
insert more than 200 characters into the `geocoded_address` field. The value of
this field comes from the `formatted_address` field in the Google geocoder
response. We could not find any documentation mentioning how long this field
might be so we took a guess and made the field 5x larger that it was.