Handle non-unicode file uploads
Previously this would cause a UnicodeDecodeError and fail. Now we
send back an error informing the user about the cause of the
error. We also record this in Rollbar. User identifying details
are kept in Rollbar occurrences, rather than Rollbar items, so
that they can be removed for GDPR if needed.