Use a different marker for selected facility
Update the FacilitiesMap component to update the marker icon used for
the currently selected marker, as detected by the presence of an OAR ID
in the route match.

When there are multiple other markers at the same point as the selected
marker, update their markers to be selected, too, so that clicking
through the popup menu continues displaying the selected marker.

Update `withQueryStringSync` to prevent automatically loading a full set
of facilities data along with the details facility when the user arrives
directly at the application via a route like
`/facilities/OAR-ID?name=shoe`. This was an unanticipated behavior that
added a lot of complexity to the selected marker behavior. The downside
is that sharing `/facilities/OAR-D?name=shoe` will now only load the
single facility's details and not load all the rest of the facilities.