Fix My Facilities navigation bug
Previously clicking a "My Facilities" link would not work as expected
because the `?contributors=id` querystring would get vaporized by the
`withQueryStringSync` component before it became a filter in the the
Redux store. Since there were two links -- one from the list page, the
other from the navbar dropdown -- there are two fixes here:

First, for the lists page link I removed a condition from
`withQueryStringSync` that would skip hydrating from the querystring in
`componentDidMount` if there were already some filters in the app.

Secondly, for the navbar dropdown I replaced the `Link` element with a
call to `setFiltersFromQueryString`, giving it a custom querystring
parameter, and an ensuring call to `fetchFacilities`.