Use raw SQL to append processing results with AWS Batch job IDs
For a list with 4000 items, we were running 12,000 individual update queries
just to append the AWS Batch job IDs to the `processing_results`. If we use
raw SQL we can drop from 12,000 queries down to 3.

The tradeoffs are breaking out of the ORM, which we mitigate by using `_meta` to
dynamically fetch the table name, and no longer writing the job ID with a row
index suffix the way it appears in the AWS Batch console, which is not really
required because we have access to the `row_id` and we are adding an additional
`is_array` flag to the processing result.