Send back list of distinct statuses for all items
Instead of sending back a single status and doing the aggregation
on the server side, it is better to send back a list of all the
statuses and let the client side decide how to present it.

Also, instead of creating another layer of enumerated statuses,
simply use the existing statuses to keep things simple.