Fix proptypes error
In #379 I accidentally introduced a react-router related prop-types
error, caused due to what I think is quirky incompatibility between
React.memo and the version of react-router we're using here: an error
was being logged to indicate that the RouteNotFound component was
getting passed as an object rather than a function to the `component`
prop.

To fix this I adjusted the route match to use `render={() => <Component
/>}` which turns it into an explicit function call. Since it's memoized
it should still be cached.