Autofocus first user input on forms
Connects #273

This also disables a linting rule against using the autoFocus prop for
the single specific line in which it needs to be turned off. Since the
inputs we are focusing will always be at the top of the page -- in order
to replicate standard HTML form behavior -- this seems to be okay.

See also
https://w3c.github.io/html/sec-forms.html#autofocusing-a-form-control-the-autofocus-attribute

In this case we are using the prop to focus inputs at the top of the
viewport.