Demo Use
Shows how the variable will be used in production. To demo this,
add REACT_APP_GOOGLE_ANALYTICS_KEY to your `.env.backend` file.

Run `cibuild` and `server`, then check in :6543/ and :8081/
that the variable is logged to the console.

This content will be replaced with actual Google Tag Manager code.