Use params dict rather than a template string for geocode requests
An unescapped # character in the address was causing geocodes to fail. To avoid
this and other problems with URL encoding we are now making use of the `params`
argument to `requests.get`.

Switching to using params broke the
`test_ungeocodable_address_returns_zero_resusts` test because it was now
returning a suggestion. We changed it to test a more un-geocodable value.