Update summary status on each Confirm / Reject
Now, for every Confirm / Reject action, we send back an
updated list of statuses. This list is used to update the
summary message atop the table, so on each interaction by
the user the status is accurate and up to date.

This does mean that the shape of the data from the confirm
and reject endpoints is different from the other list item
details endpoints, but perhaps that is okay for now.