Rework reprocess_geocode_failures to use AWS Batch
The original version of this management command ran all the processing in the
context of the command itself. This worked correctly for test data, but the
process was killed when trying to run it on production-sized data.

The batch processing tasks are designed to only work on items in a specific
status so we can safely change the status on just the items we need to process
and then resubmit the entire list to the geocoding and match tasks.

We have added a `skip_parse` option to `submit_jobs` as a convenience since we
know that we will not have to re-parse any items.