Add encrypted RDS database and supporting security groups
Create an additional RDS database with encrypted storage. Include all of
the supporting security group resources to facilitate communication to
it.

In addition, downgrade the Terraform RDS module to a version that does
not support enhanced monitoring (the IAM role names collide) and mark
all of the prior security groups for deletion.