Download facility list items one page at a time
Downloading multiple pages in parallel was overloading the service in staging.
To decrease the load we have rewritten the page request loop to request pages of
results in series.

We use a traditional for loop because eslint discourages using for/of but we
want to explicitly await in each iteration of the loop rather than create async
functions with map or forEach.

We disable `no-await-in-loop` because we intentionally want to run our items one
at a time in series.