Add support for filtering list items by NEW_FACILITY meta state
NEW_FACILITY is not a real workflow state. A `FacilityListItem` is considered a
NEW_FACILITY when it has a bidirectional relationship with a `Facility`. We want
to avoid showing NEW_FACILITY items when filtering by the MATCHED or
CONFIRMED_MATCH statuses. To handle all of this we create a set of special case
filter statements and compose them together with `status=` statements for the
other statuses.