Refresh facility list items when filtering by POTENTIAL_MATCH
If you have filtered the list of facility list items by `POTENTIAL_MATCH` and
then confirm or reject an item we need to update the current page since the
count of items in the `POTENTIAL_MATCH` status is now different.

To accomplish this we convert the FacilityListItemsTable from a functional to a
class component so we can add logic to `componentDidUpdate` that conditionally
triggers a refresh.