Use some read only fields in the Django admin
The Facility, FacilityListItem, and FacilityMatch admin were loading combo boxes
with all possible related objects. With a database populated with 10s of
thousands of records this would make admin page requests so slow that the ECS
task would be terminated.

This commit fixes that by simply making the related fields read only. This works
because we would not want to adjust these specific foreign key relationships
anyway.

While making some fields read only we also took the opportunity to also make the
JSON status fields read only and add a helper to render them in a more readable
way.