Show facility list details the contributor profile page
We currently allow contributors to specify a detailed description for a list but
prior to this commit we were not displaying that description publicly. From the
facility detail page we link to the profile page of any contributors associated
with that facility so the profile page is a suitable place to display the
facility lists and their descriptions.

Contributors have a dedicated UI for viewing their own lists so we are only
showing the list details on public profile pages, not the profile page of the
logged in contributor.