Fetch client country code and write to app state
We need to detect if someone is browsing the app from China so that we can
change the way in which we load our map components. To facilitate this we make a
request to the free userinfo.io service when the `App` component loads and save
the returned country to the app state.