Add search control to list page
This control uses the new endpoint to get a filtered list
of items where only those that match the search term in
their name or address show. The control supports mouse and
keyboard inputs. The search can also be shared across the
web via the URL, and initializes to the right content when
reloading the page.