Load Google translate from the proper host based on country
Loading Google translate library in China requires that we request the script
from google.cn rather than google.com. To allow this conditional loading we
moved the script tag from `index.html` into a new component that uses a
lifecycle method to append the script tag with the proper `src`.