Add verification fields and history tracking to Contributor
We want to allow superusers to mark certain brand accounts as verified. This
implements the simplest version of that feature, adding a verified flag and a
description box accessible via the Django Admin. We opt the model into history
tracking so that we have a log of changes to the verified status.