Show verified status on the user profile page
Adds an is_verified field to the profile serializer and related front end code
to show an icon on the profile page when the contributor is verified.

We have opted to show the name of the contributor as the title rather than the
static "Profile" to emphasize what the page is about, to save some vertical
space, and to provide a high-profile place on the page to show the verification
badge.