Show verified badge on facility detail page
Showing the verified badge next to a contributor on the facility details page
requires us to have the verified status of contributors available in the
facility details API response. The previous implementation of the
`Facility.contributors` model method was returning a string representation of
the contributor. To better separate concerns, this helper now returns a list of
`FacilityList` objects that contributed the Facility, leaving the rendering up
to the serializer. We are now returning a list of objects that include an
`is_verified` flag that we use to conditionally render a verified SVG icon.

In order to reuse the same icon we used on the profile page, we are now passing
through the `style` prop from `BadgeVerified` to the internal `SvgIcon` in order
to adjust the size, margins, etc.