Add front-end for Dashboard Lists view
Adds actions, reducers, components for Dashboard List,
a view where superusers can see all the lists of all the
contributors.

They are presented with a dropdown to select a contributor,
selecting one in which triggers a fetch of the lists by
that contributor. The lists are ordered by recency. The
table looks similar to that of one's own list, except for
the additional "Date Created" column. Clicking any row
takes the user to a "read-only" view of the list.

The selected contributor is added to the URL for sharing
and refreshing purposes. Coming back from a list detail
view resumes the dashboard lists view in the state it
was left in.