Show a list of claimed facilities
Repurposes the "My Facilities" link the navbar to show a page listing all
approved facility claims. Items in the list link to the facility profile page
where additional information about the facility can be collected.

The "My Facilities" link now depends on the claim a facility feature so we have
updated the `NavbarLoginButtonGroup` to only show the link when the feature flag
is enabled.