Add .prettierrc.json config file
Add a .prettierrc.json configuration file with rules which are
compatible with our existing set of ESLint rules. This enables
developers to run Prettier manually in their editors on new files while
working, saving the burden of having to manually format the code.

I intentionally did not set this up to run prettier on file saves,
although that is theoretically possible now that we are using
react-app-rewired. Setting it up to run on saves would cause a large
initial changeset because many of the project's files have not been
formatted with Prettier.

Closes #591