Add FacilityAlias model
This tables allows us to keep track of the OAR IDs for facilities that are
merged with other facilities or deleted so that we can report on their removal
and redirect to their replacement facilities.

We applied the same primary key field settings to `oar_id` that were used for
`Facility.id` because we only want a single alias for each OAR ID and we do not
want it to change.