Add a stub destroy action to the FacilitiesViewSet
To allow for deletes we need to convert from using `ReadOnlyModelViewSet` to
manually opting in to the list, retrieve, and delete methods.

We will be reserving this action for superusers, so we exclude it from the
public API documentation by defining a custom `AutoSchema` subclass and
overriding the appropriate method.