Implement facility merge endpoint
We merge a facility by:
- Updating all the `FacilityMatch` records and their related `FacilityListItem`s
  to point to the the target `Facility`
- Changing the `FacilityMatch status to `MERGED` and adding a processing result
  item to the `FacilityListItem` with a `MERGE_FACILITY` action
- Creating a `FacilityAlias` with a type of `MERGE` pointing the merged OAR ID
  to the target facility
- Deleting the merged facility record.

We needed to update the filters used by the `Facility` serialization process so
that `MERGED` facilities would show up properly in on the facility details page.

After merging we return the details of the "target" `Facility` to make updating
the UI with the results of the merge easier.

Now that multiple items from the same `FailityList` can be merged under one
`Facility` we need to make sure and return a distinct list of `FacilityList`
from the facility details endpoint to avoid displaying duplicate links.

FIXUP return facility details after successful merge

FIXUP return distinct items from Facility.contributors()