Add OAR_CLIENT_KEY
The OAR_CLIENT_KEY is set in an environment variable at provision time and is
made available as a Django setting and on the client app `ENVIRONMENT` object.
The client app can pass this key along with API requests and the server can
verify it.

We need to provide stakeholders with metrics on programmatic API usage, which
means that we need to start requiring a token for all API requests. We still,
however, need to allow anonymous access to the search APIs via the web
application. The client key will be used to protect anonymous access to search
APIs from the web application without requiring a "dummy" user account.

We set an empty client key environment variable in docker-compose.yml so we can
more easily test endpoints in development.