Authorize anonymous web client API requests using OAR_CLIENT_KEY
We need to provide stakeholders with metrics on programmatic API usage, which
means that we need to require a token for all API requests. We still, however,
need to allow anonymous access to the search APIs via the web application.

We implement this by passing a client key will all API requests made by the web
client and replacing the default django-rest-framework `AllowAny` permission
with a custom permission class that checks for this client key if the request is
not otherwise authorized.

We took advantage of the fact that all request from the client are made via our
`csrfRequest` "wrapper" around `axios` and added the key handling to it. This
prevents the need to make any modifications to existing code and allows
application actions to continue to make requests without worrying about
authorization details.