Add middleware to log requests made with token authentication
We create a `RequestLog` row for all requests that use token-based
authentication so we can report on non-webapp API usage.

We do not want a failure in the request logging to disrupt the handling of a
response to we catch all exceptions and send them to Rollbar, failing silently
if the call to Rollbar, in turn, raises an exception.