Add changed values to facility update emails
When notifying contributors when facility claimants change the profile of a
facility we want to include the specific public details that were changed.

The non-standard position of the `endfor` in the plain text email template is
intentional and ensures that extra blank lines are not included.

When capitalizing the verbose name, we are manually uppercasing the first letter
because Python's built-in `capitalize` method changes all characters but the
first to lower case, which is not what we want in the case of a verbose name
like "facility name in English."