Display RouteNotFound on unmatched routes
Adjust react router switch statements to display the RouteNotFound
component on unmatched routes. For routes which are leaf nodes, add an
`exact` prop to these routes to indicate that they should only match
exactly; for matches that are not, adjust order of cases in the `Switch`
blocks to ensure routes are rendered appropriately.