Add facility product and production type to form and facility detail
A claimant can enter arbitrary product or production types, so we use the
`Creatable` version of `react-select` to collect them on the profile form. To
encourage reuse we populate the select boxes with a combination of a "seed" list
and all the arbitrary values that have been added across all claims.