Add admin reports
Provides a quick, simple way to do ad-hoc reporting via raw SQL queries.

We add a pair of new admin views so that the reports are only accessible to
superusers.

A new report can be added just by adding a `.sql` file to `/api/reports/`. The
name of the report is derived from the name of the `.sql` file.

We return and display the raw SQL along with the results to make easy to debug
and answer questions about how the report was generated.

We opt for raw SQL for the flexability it provides. We will never be writing
data so the ORM would cause more friction than provide benefit.