Wrap the facility search results list with react-infinite
We plan to allow for the display of all matching search results rather than
capping the result count at 500. Using `react-infinite` will allow us to trigger
a request for additional items when the person searching for facilities scrolls
to the bottom of the partial list.

`react-infinite` forces us to explicitly set the size of the container and the
size of each item. We are temporarily using a hardcoded 120px height for list
items which looks ok for most results. In a future commit we can develop a
"height calculator" that would allow us to resume using variable height list
items.