Add react-infinite-any-height
`react-infinite-any-height` is a wrapper around `react-infinite` that wraps each
list item with a span and uses that span to calculate the height, preventing the
need to force all list items to be the same height. There is a lot of
variability in the length of facility names and addresses. On narrow screens
some list items can be twice as tall as others.