Add a cacheKey for tile requests
Add a cacheKey to the tile requests, which originates from the max
updated_at value on the facilities table. Also add infrastructure to
get the cacheKey for the vector tile layer:

- initially from window.ENVIRONMENT.TILE_CACHE_KEY
- subsequently, from a tile cache key endpoint consumed prior to making
a facilities request

Also update the vector tile layer component to throw an error if cache
key is ever unset, and the tile endpoint to return a 400 if no cache key
is sent.