Set map view on facility contextually
When the user has arrived at the map by visiting a URL which contains an
OAR ID, set the map view to center on the facility when the facility
data returns. Otherwise, selecting a facility will not re-center the
viewport.

If the OAR ID encoded in a URL returns an error, switch off the setView
behavior to ensure that the next manually selected facility does not
incite a setView call.

If a facility which is off-screen is selected (as may happen when
selecting a facility from the list sidebar), center the map view on the
facility when its data returns.