Add tiler method for getting facility grid vector tiles
We create a fishnet grid tile by using `mercantile` to generate the 16 square
polygons that fit within the tile and building a query that makes a literal
table containing those 16 squares. We left join that literal table to the
results of the filtered facility query, aggregating to get the count of
facilities for each cell.

We generate both a lat/lng polygon and an MVT polygon for each row of our
literal fishnet table so we can use the lat/lng geometry for filtering the
features and the MVT geometry for rendering.

If the textual size of these queries becomes a performance issue, we can look
into moving the fishnet creation logic into an PostgreSQL function.