Show facility fishnet grid layer on map
We add a new layer based on the existing VectorTileFacilitiesLayer and add
support to both layers for setting `minZoom` and `maxZoom`. By trial and error
we chose 11/12 as the zoom level at which we transition from the grid layer to
the individual marker layer.

The color ramp was generated using blues used in the app already as endpoints.
We incorporated a z-dependent scaling of the color ramp so that at higher zoom
levels where there are fewer facilities per cell we make more use of the
available range of colors.

The count thresholds between colors were also chosen by trial and error and will
likely need to be adjusted to effectively represent production data.