Enable downloading all facilities as CSV
Using the machinery implemented for the infinite scroll sidebar, enable
signed-in users to download a CSV containing all of the facilities
matching any given search params, including all the OAR's facility data
when no search params are present.

Also increase the size of the facilities pages from 25 to 100. We want
to keep this value shared between the infinite scroll and the CSV
download so that we can reuse the existing machinery and so that after
downloading a CSV, the user can scroll through all the facilities in the
sidebar and can download the CSV again without undertaking new requests.

The change also adds a loading indicator to display in place of the
download CSV button while a download is underway.